Release 1.0.3
This release adds the binary "ping-multi"
which is a replacement of https://github.com/famzah/ping-multi

This "ping-multi" supports IPv6 but uses and parses
the external "ping" command.

The binary "ping-geo" is removed and instead remote hosts
can easily be specified by "pinghost@ssh_host.example.com"
directly in "ping-multi".

A direct command-line interface is provided by
the binary "ping-raw-multi" for advanced usage not covered
by the defaults in "ping-multi".